Fix issue where audio output is not updated if chosen before meeting start
