+++ docs/interfaces/mediastreambrokerobserver.html
@@ -85,6 +85,7 @@
 			<section class="tsd-panel">
 				<h3>Implemented by</h3>
 				<ul class="tsd-hierarchy">
+					<li><a href="../classes/defaultaudiomixcontroller.html" class="tsd-signature-type">DefaultAudioMixController</a></li>
 					<li><a href="../classes/defaultaudiovideocontroller.html" class="tsd-signature-type">DefaultAudioVideoController</a></li>
 					<li><a href="../classes/noopaudiovideocontroller.html" class="tsd-signature-type">NoOpAudioVideoController</a></li>
 				</ul>
